Extract IStringFmt.js from IString.js
Move formatting methods (format, _testChoice, _isIntlPluralAvailable,
formatChoice, setLocale, getLocale) out of IString.js into a standalone
IStringFmt.js module, mixed back onto IString.prototype via
Object.assign. This completes Phase 2 of the split described in
docs/IString-Refactoring.md (IString.js: 1158 -> 712 lines).

- formatChoice's `new IString(...)` calls become `new this.constructor(...)`
  so IStringFmt.js never requires IString.js directly (no circular dep)
- IString.prototype.constructor is restored explicitly, since replacing
  .prototype with an object literal drops it, which broke the
  this.constructor calls above
- _testChoice/setLocale now call PluralUtils directly instead of
  reaching through the IString._fncs/loadPlurals aliases
- Add js/test/root/testistringfmt.js covering format/formatChoice/
  setLocale/getLocale end-to-end

No functional changes intended; several pre-existing bugs in the moved
code (found during review) are left untouched for a follow-up commit.
